+++ packages/scrolly-kit/src/lib/index.ts
@@ -14,7 +14,6 @@ export { default as ScrollIndicator } from './components/ScrollIndicator.svelte'
 export { default as CopyCodeBlock } from './components/CopyCodeBlock.svelte';
 export { default as CodeExplainer } from './components/CodeExplainer.svelte';
 export { default as Spinner } from './components/Spinner.svelte';
-export { default as Meta } from './components/Meta.svelte';
 export { default as StoryHeader } from './components/StoryHeader.svelte';
 export { default as Nav } from './components/Nav.svelte';
 export { default as NavMenu } from './components/NavMenu.svelte';
@@ -50,7 +49,6 @@ export type {
   RangeSliderProps,
   ScrollIndicatorProps,
   SpinnerProps,
-  MetaProps,
   RenderContentProps,
   CodeStep,
   CodeExplainerData,